#34 - 로깅 미들웨어 수정
