+++ pg_rest_api/settings/components/middlewares.py
@@ -6,6 +6,6 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    'pg_rest_api.middleware.LoggingMiddleware',
-    'pg_rest_api.middleware.ResponseFormattingMiddleware'
+    'pg_rest_api.middleware.logging.LoggingMiddleware',
+    'pg_rest_api.middleware.response_formatting.ResponseFormattingMiddleware'
 ]
